Add health-check endpoint
Sometimes it is good to have an endpoint available which will return
HTTP 200 when the service is up and running, for example for load
balancers or caches. Rib currently only offers HTTP 200 when POSTing,
which not every software supports when sending health-check pings or
when retrieving metrics, but those are behind HTTP Basic auth which,
again, not every software supports.

This adds an endpoint which returns HTTP 200 on GET requests without any
fuss. It even returns a trailing newline so curl users don't get a weird
prompt on bash.